implementation/requirements.txt

langgraph
langchain-core
langchain-openai
openai
python-dotenv
IPython
matplotlib
pandas
numpy